Cleanup of weight channels and fir modules
Fixed the naming conventions of these modules so that they adhere to
PEP8. Broke out the actual computations into a function that can be used
independently of the module (facilitates testing and composition of new
modules such as those required for Brad's RDT so he can reuse the
computations without having to cut-and-paste the module code itself).
The new functions require the input data and the coefficients. From now
on, the class method `my_eval` only has one job. Obtain the coefficients
and pass them to the actual evaluation function.