Revamped phi system
Modules are now responsible for returning a dictionary for phi. Each
entry in the dictionary is the coefficient name and the value is the
initial guess for that coefficient. Stacks will assemble the
dictionaries returned by each module into a list. This list is then
provided to the fitter. The fitter is then responsible for packing this
list of dictionaries into a vector and then unpacking the vector into a
list of dictionaries. Helper functions (with doctests) are provided for
this purpose.