Updated signal loading to support Numpy binary
Loading large signals from CSV can be computationally slow (in the order
of seconds). For iterative programming, we need faster loading so we can
test our code. This now supports loading from Numpy binary files (and if
it loads from CSV, you can tell it to cache the result to a Numpy binary
file for faster future loads).