Updates to the modules
Removed the get_inputs and get_outputs. These were originally proposed
in the mockup, but we aren't currently using them. Can be added in at a
later step if needed.

Updated modules with the ability to generate tensor symbolic operations
that can be used for fitting with libraries such as PyMC3 (allows for
Bayes and variational inference steps).